Change RealmResourcesCache name to InitialDataDownloaderResourcesSync
