+++ godtools/App/Services/InitialDataDownloader/InitialDeviceResourcesLoader.swift
@@ -15,19 +15,19 @@ class InitialDeviceResourcesLoader {
     private let legacyRealmMigration: LegacyRealmMigration
     private let attachmentsFileCache: AttachmentsFileCache
     private let translationsFileCache: TranslationsFileCache
-    private let realmResourcesCache: RealmResourcesCache
+    private let resourcesSync: InitialDataDownloaderResourcesSync
     private let favoritedResourcesCache: FavoritedResourcesCache
     private let languagesCache: RealmLanguagesCache
     private let deviceLanguage: DeviceLanguageType
     private let languageSettingsCache: LanguageSettingsCacheType
         
-    required init(realmDatabase: RealmDatabase, legacyRealmMigration: LegacyRealmMigration, attachmentsFileCache: AttachmentsFileCache, translationsFileCache: TranslationsFileCache, realmResourcesCache: RealmResourcesCache, favoritedResourcesCache: FavoritedResourcesCache, languagesCache: RealmLanguagesCache, deviceLanguage: DeviceLanguageType, languageSettingsCache: LanguageSettingsCacheType) {
+    required init(realmDatabase: RealmDatabase, legacyRealmMigration: LegacyRealmMigration, attachmentsFileCache: AttachmentsFileCache, translationsFileCache: TranslationsFileCache, resourcesSync: InitialDataDownloaderResourcesSync, favoritedResourcesCache: FavoritedResourcesCache, languagesCache: RealmLanguagesCache, deviceLanguage: DeviceLanguageType, languageSettingsCache: LanguageSettingsCacheType) {
         
         self.realmDatabase = realmDatabase
         self.legacyRealmMigration = legacyRealmMigration
         self.attachmentsFileCache = attachmentsFileCache
         self.translationsFileCache = translationsFileCache
-        self.realmResourcesCache = realmResourcesCache
+        self.resourcesSync = resourcesSync
         self.favoritedResourcesCache = favoritedResourcesCache
         self.languagesCache = languagesCache
         self.deviceLanguage = deviceLanguage
@@ -36,7 +36,7 @@ class InitialDeviceResourcesLoader {
     
     func loadAndCacheInitialDeviceResourcesIfNeeded(completeOnMain: @escaping (() -> Void)) {
         
-        guard !realmResourcesCache.resourcesAvailable else {
+        guard !resourcesSync.resourcesAvailable else {
             completeOnMain()
             return
         }
@@ -60,7 +60,7 @@ class InitialDeviceResourcesLoader {
     
     private func loadAndCacheLanguagesPlusResourcesPlusLatestAttachmentsAndTranslations(complete: @escaping ((_ result: Result<ResourcesCacheResult, Error>?) -> Void)) {
         
-        let realmResourcesCache: RealmResourcesCache = self.realmResourcesCache
+        let resourcesSync: InitialDataDownloaderResourcesSync = self.resourcesSync
         
         realmDatabase.background { (realm: Realm) in
             
@@ -79,7 +79,7 @@ class InitialDeviceResourcesLoader {
                     resourcesPlusLatestTranslationsAndAttachments: resources
                 )
                 
-                let result: Result<ResourcesCacheResult, Error> = realmResourcesCache.cacheResources(realm: realm, downloaderResult: downloaderResult)
+                let result: Result<ResourcesCacheResult, Error> = resourcesSync.cacheResources(realm: realm, downloaderResult: downloaderResult)
                 
                 complete(result)
             }
